Initial version_checker refactor
