+++ medusa/updater/github_updater.py
@@ -1,0 +1,399 @@
+# coding=utf-8
+
+from __future__ import unicode_literals
+
+import logging
+import os
+import platform
+import re
+import subprocess
+from builtins import str
+
+from medusa import app, notifiers
+from medusa.logger.adapters.style import BraceAdapter
+from medusa.updater.update_manager import UpdateManager
+
+
+ERROR_MESSAGE = ('Unable to find your git executable. Set git executable path in Advanced Settings '
+                 'OR shutdown application and delete your .git folder and run from source to enable updates.')
+
+log = BraceAdapter(logging.getLogger(__name__))
+log.logger.addHandler(logging.NullHandler())
+
+
+class GitUpdateManager(UpdateManager):
+    def __init__(self):
+        super(GitUpdateManager, self).__init__()
+        self._git_path = self._find_working_git()
+        self.github_org = self.get_github_org()
+        self.github_repo = self.get_github_repo()
+
+        self.branch = app.BRANCH = self._find_installed_branch()
+
+        self._cur_commit_hash = None
+        self._newest_commit_hash = None
+        self._num_commits_behind = 0
+        self._num_commits_ahead = 0
+        self._cur_version = ''
+
+    def get_cur_commit_hash(self):
+        return self._cur_commit_hash
+
+    def get_newest_commit_hash(self):
+        return self._newest_commit_hash
+
+    def get_cur_version(self):
+        if self._cur_commit_hash or self._find_installed_version():
+            self._cur_version = self._run_git(self._git_path, 'describe --tags --abbrev=0 {0}'.format(self._cur_commit_hash))[0]
+        return self._cur_version
+
+    def get_newest_version(self):
+        if self._newest_commit_hash:
+            self._cur_version = self._run_git(self._git_path, 'describe --tags --abbrev=0 ' + self._newest_commit_hash)[0]
+        else:
+            self._cur_version = self._run_git(self._git_path, 'describe --tags --abbrev=0 ' + self._cur_commit_hash)[0]
+        return self._cur_version
+
+    def get_num_commits_behind(self):
+        return self._num_commits_behind
+
+    def get_num_commits_ahead(self):
+        return self._num_commits_ahead
+
+    def _find_working_git(self):
+        test_cmd = 'version'
+
+        if app.GIT_PATH:
+            main_git = '"' + app.GIT_PATH + '"'
+        else:
+            main_git = 'git'
+
+        log.debug(u'Checking if we can use git commands: {0} {1}', main_git, test_cmd)
+        _, _, exit_status = self._run_git(main_git, test_cmd)
+
+        if exit_status == 0:
+            log.debug(u'Using: {0}', main_git)
+            return main_git
+        else:
+            log.debug(u'Not using: {0}', main_git)
+
+        # trying alternatives
+
+        alternative_git = []
+
+        # osx people who start sr from launchd have a broken path, so try a hail-mary attempt for them
+        if platform.system().lower() == 'darwin':
+            alternative_git.append('/usr/local/git/bin/git')
+
+        if platform.system().lower() == 'windows':
+            if main_git != main_git.lower():
+                alternative_git.append(main_git.lower())
+
+        if alternative_git:
+            log.debug(u'Trying known alternative git locations')
+
+            for cur_git in alternative_git:
+                log.debug(u'Checking if we can use git commands: {0} {1}', cur_git, test_cmd)
+                _, _, exit_status = self._run_git(cur_git, test_cmd)
+
+                if exit_status == 0:
+                    log.debug(u'Using: {0}', cur_git)
+                    return cur_git
+                else:
+                    log.debug(u'Not using: {0}', cur_git)
+
+        # Still haven't found a working git
+        # Warn user only if he has version check enabled
+        if app.VERSION_NOTIFY:
+            app.NEWEST_VERSION_STRING = ERROR_MESSAGE
+
+        return None
+
+    @staticmethod
+    def _run_git(git_path, args):
+
+        output = err = exit_status = None
+
+        if not git_path:
+            log.warning(u"No git specified, can't use git commands")
+            app.NEWEST_VERSION_STRING = ERROR_MESSAGE
+            exit_status = 1
+            return output, err, exit_status
+
+        # If we have a valid git remove the git warning
+        # String will be updated as soon we check github
+        app.NEWEST_VERSION_STRING = None
+
+        cmd = git_path + ' ' + args
+
+        try:
+            log.debug(u'Executing {cmd} with your shell in {dir}', {'cmd': cmd, 'dir': app.PROG_DIR})
+            p = subprocess.Popen(cmd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
+                                 shell=True, cwd=app.PROG_DIR)
+            output, err = p.communicate()
+            exit_status = p.returncode
+
+            # Convert bytes to string in python3
+            if isinstance(output, (bytes, bytearray)):
+                output = output.decode('utf-8')
+
+            if output:
+                output = output.strip()
+
+        except OSError:
+            log.info(u"Command {cmd} didn't work", {'cmd': cmd})
+            exit_status = 1
+
+        if exit_status == 0:
+            log.debug(u'{cmd} : returned successful', {'cmd': cmd})
+            exit_status = 0
+
+        elif exit_status == 1:
+            if output:
+                if 'stash' in output:
+                    log.warning(u"Enable 'git reset' in settings or stash your changes in local files")
+                else:
+                    log.warning(u'{cmd} returned : {output}', {'cmd': cmd, 'output': output})
+            else:
+                log.warning(u'{cmd} returned no data', {'cmd': cmd})
+            exit_status = 1
+
+        elif exit_status == 128 or 'fatal:' in output or err:
+            log.warning(u'{cmd} returned : {output}', {'cmd': cmd, 'output': output})
+            exit_status = 128
+
+        else:
+            log.warning(u'{cmd} returned : {output}. Treat as error for now', {'cmd': cmd, 'output': output})
+            exit_status = 1
+
+        return output, err, exit_status
+
+    def _find_installed_version(self):
+        """Attempt to find the currently installed version of the application.
+
+        Uses git show to get commit version.
+
+        :return: True for success or False for failure
+        """
+        output, _, exit_status = self._run_git(self._git_path, 'rev-parse HEAD')  # @UnusedVariable
+
+        if exit_status == 0 and output:
+            cur_commit_hash = output.strip()
+            if not re.match('^[a-z0-9]+$', cur_commit_hash):
+                log.warning(u"Output doesn't look like a hash, not using it")
+                return False
+            self._cur_commit_hash = cur_commit_hash
+            app.CUR_COMMIT_HASH = str(cur_commit_hash)
+            return True
+        else:
+            return False
+
+    def _find_installed_branch(self):
+        branch_info, _, exit_status = self._run_git(self._git_path, 'symbolic-ref -q HEAD')  # @UnusedVariable
+        if exit_status == 0 and branch_info:
+            branch = branch_info.strip().replace('refs/heads/', '', 1)
+            if branch:
+                app.BRANCH = branch
+                return branch
+        return ''
+
+    def _check_github_for_update(self):
+        """
+        Uses git commands to check if there is a newer version that the provided
+        commit hash. If there is a newer version it sets _num_commits_behind.
+        """
+
+        self._num_commits_behind = 0
+        self._num_commits_ahead = 0
+
+        # update remote origin url
+        self.update_remote_origin()
+
+        # get all new info from github
+        output, _, exit_status = self._run_git(self._git_path, 'fetch --prune %s' % app.GIT_REMOTE)
+        if not exit_status == 0:
+            log.warning(u"Unable to contact github, can't check for update")
+            return
+
+        # get latest commit_hash from remote
+        output, _, exit_status = self._run_git(self._git_path, 'rev-parse --verify --quiet "@{upstream}"')
+
+        if exit_status == 0 and output:
+            cur_commit_hash = output.strip()
+
+            if not re.match('^[a-z0-9]+$', cur_commit_hash):
+                log.debug(u"Output doesn't look like a hash, not using it")
+                return
+
+            else:
+                self._newest_commit_hash = cur_commit_hash
+        else:
+            log.debug(u"git didn't return newest commit hash")
+            return
+
+        # get number of commits behind and ahead (option --count not supported git < 1.7.2)
+        output, _, exit_status = self._run_git(self._git_path, 'rev-list --left-right "@{upstream}"...HEAD')
+        if exit_status == 0 and output:
+
+            try:
+                self._num_commits_behind = int(output.count('<'))
+                self._num_commits_ahead = int(output.count('>'))
+
+            except Exception:
+                log.debug(u"git didn't return numbers for behind and ahead, not using it")
+                return
+
+        log.debug(u'cur_commit = {0}, newest_commit = {1}, num_commits_behind = {2}, num_commits_ahead = {3}',
+                  self._cur_commit_hash, self._newest_commit_hash, self._num_commits_behind, self._num_commits_ahead)
+
+    def set_newest_text(self):
+        # if we're up to date then don't set this
+        app.NEWEST_VERSION_STRING = None
+
+        if self._num_commits_behind > 0 or self._is_hard_reset_allowed():
+
+            base_url = 'http://github.com/' + self.github_org + '/' + self.github_repo
+            if self._newest_commit_hash:
+                url = base_url + '/compare/' + self._cur_commit_hash + '...' + self._newest_commit_hash
+            else:
+                url = base_url + '/commits/'
+
+            newest_text = 'There is a <a href="' + url + '" onclick="window.open(this.href); return false;">newer version available</a> '
+            newest_text += " (you're " + str(self._num_commits_behind) + ' commit'
+            if self._num_commits_behind > 1:
+                newest_text += 's'
+            newest_text += ' behind'
+            if self._num_commits_ahead > 0:
+                newest_text += ' and {ahead} commit{s} ahead'.format(ahead=self._num_commits_ahead,
+                                                                     s='s' if self._num_commits_ahead > 1 else '')
+            newest_text += ') &mdash; <a href="' + self.get_update_url() + '">Update Now</a>'
+
+        elif self._num_commits_ahead > 0:
+            newest_text = u'Local branch is ahead of {0}. Automatic update not possible'.format(self.branch)
+            log.warning(newest_text)
+
+        else:
+            return
+
+        app.NEWEST_VERSION_STRING = newest_text
+
+    def need_update(self):
+
+        if self.branch != self._find_installed_branch():
+            log.debug(u'Branch checkout: {0}->{1}', self._find_installed_branch(), self.branch)
+            return True
+
+        self._find_installed_version()
+        if not self._cur_commit_hash:
+            return True
+        else:
+            try:
+                self._check_github_for_update()
+            except Exception as e:
+                log.warning(u"Unable to contact github, can't check for update: {0!r}", e)
+                return False
+
+            if self._num_commits_behind > 0 or self._num_commits_ahead > 0:
+                return True
+
+        return False
+
+    def can_update(self):
+        """Return whether update can be executed.
+
+        :return:
+        :rtype: bool
+        """
+        return self._num_commits_ahead <= 0 or self._is_hard_reset_allowed()
+
+    def update(self):
+        """Call git pull origin <branch> in order to update the application.
+
+        Returns a bool depending on the call's success.
+        """
+        # update remote origin url
+        self.update_remote_origin()
+
+        # remove untracked files and performs a hard reset on git branch to avoid update issues
+        if self._is_hard_reset_allowed():
+            self.reset()
+
+        # Executing git clean before updating
+        self.clean()
+
+        if self.branch == self._find_installed_branch():
+            _, _, exit_status = self._run_git(self._git_path, 'pull -f %s %s' % (app.GIT_REMOTE, self.branch))  # @UnusedVariable
+        else:
+            _, _, exit_status = self._run_git(self._git_path, 'checkout -f ' + self.branch)  # @UnusedVariable
+
+        # Executing git clean after updating
+        self.clean()
+
+        if exit_status == 0:
+            self._find_installed_version()
+            # Notify update successful
+            if app.NOTIFY_ON_UPDATE:
+                try:
+                    notifiers.notify_git_update(app.CUR_COMMIT_HASH or '')
+                except Exception:
+                    log.debug(u'Unable to send update notification. Continuing the update process')
+            return True
+
+        else:
+            return False
+
+    @staticmethod
+    def _is_hard_reset_allowed():
+        """Return whether git hard reset is allowed or not.
+
+        :return:
+        :rtype: bool
+        """
+        return app.GIT_RESET and (not app.GIT_RESET_BRANCHES or
+                                  app.BRANCH in app.GIT_RESET_BRANCHES)
+
+    def clean(self):
+        """Call git clean to remove all untracked files.
+
+        It only affects source folders and libX and extX folders,
+        to prevent deleting untracked user data not known by .gitignore
+
+        :return:
+        :rtype: int
+        """
+        # Fixes: goo.gl/tr8Awf - to be removed in the next release
+        root_dir = os.path.basename(app.PROG_DIR)
+        helper_folder = os.path.join(root_dir, 'helper')
+        helpers_folder = os.path.join(root_dir, 'helpers')
+
+        folders = (app.LIB_FOLDER, app.LIB2_FOLDER, app.LIB3_FOLDER, app.EXT_FOLDER,
+                   app.EXT2_FOLDER, app.EXT3_FOLDER, app.SRC_FOLDER, app.STATIC_FOLDER,
+                   helper_folder, helpers_folder) + app.LEGACY_SRC_FOLDERS
+        _, _, exit_status = self._run_git(self._git_path, 'clean -d -f -x {0}'.format(' '.join(folders)))
+
+        return exit_status
+
+    def reset(self):
+        """
+        Calls git reset --hard to perform a hard reset. Returns a bool depending
+        on the call's success.
+        """
+        _, _, exit_status = self._run_git(self._git_path, 'reset --hard {0}/{1}'.format
+                                          (app.GIT_REMOTE, app.BRANCH))  # @UnusedVariable
+        if exit_status == 0:
+            return True
+
+    def list_remote_branches(self):
+        # update remote origin url
+        self.update_remote_origin()
+        app.BRANCH = self._find_installed_branch()
+
+        branches, _, exit_status = self._run_git(self._git_path, 'ls-remote --heads %s' % app.GIT_REMOTE)  # @UnusedVariable
+        if exit_status == 0 and branches:
+            if branches:
+                return re.findall(r'refs/heads/(.*)', branches)
+        return []
+
+    def update_remote_origin(self):
+        self._run_git(self._git_path, 'config remote.%s.url %s' % (app.GIT_REMOTE, app.GIT_REMOTE_URL))
+        self._run_git(self._git_path, 'config remote.%s.pushurl %s' % (app.GIT_REMOTE, app.GIT_REMOTE_URL))
